+++ src/GafferScene/Instancer.cpp
@@ -264,16 +264,71 @@ struct UniqueHashPrototypeContextVariable
 	}
 };
 
+InternedString g_prototypeRootName( "root" );
+ConstInternedStringVectorDataPtr g_emptyNames = new InternedStringVectorData();
+
+struct IdData
+{
+	IdData() :
+		intElements( nullptr ), int64Elements( nullptr )
+	{
+	}
+
+	void initialize( const Primitive *primitive, const std::string &name )
+	{
+		if( const IntVectorData *intData = primitive->variableData<IntVectorData>( name ) )
+		{
+			intElements = &intData->readable();
+		}
+		else if( const Int64VectorData *int64Data = primitive->variableData<Int64VectorData>( name ) )
+		{
+			int64Elements = &int64Data->readable();
+		}
+
+	}
+
+	size_t size() const
+	{
+		if( intElements )
+		{
+			return intElements->size();
+		}
+		else if( int64Elements )
+		{
+			return int64Elements->size();
+		}
+		else
+		{
+			return 0;
+		}
+	}
+
+	int64_t element( int i ) const
+	{
+		if( intElements )
+		{
+			return (*intElements)[i];
+		}
+		else
+		{
+			return (*int64Elements)[i];
+		}
+	}
+
+	const std::vector<int> *intElements;
+	const std::vector<int64_t> *int64Elements;
+
+};
+
 // We create a seed integer that corresponds to the id by hashing the id and then modulo'ing to
 // numSeeds, to create seeds in the range 0 .. numSeeds-1 that persistently correspond to the ids,
 // with a grouping pattern that can be changed with seedScramble
-int seedForPoint( int index, const PrimitiveVariable *primVar, int numSeeds, int seedScramble )
+int seedForPoint( int index, const IdData &idData, int numSeeds, int seedScramble )
 {
-	int id = index;
-	if( primVar )
+	int64_t id = index;
+	if( idData.size() )
 	{
-		// TODO - the exception this will throw on non-int primvars may not be very clear to users
-		id = PrimitiveVariable::IndexedView<int>( *primVar )[index];
+		id = idData.element( index );
 	}
 
 	// numSeeds is set to 0 when we're just passing through the id
@@ -293,15 +348,24 @@ int seedForPoint( int index, const PrimitiveVariable *primVar, int numSeeds, int
 
 		IECore::MurmurHash seedHash;
 		seedHash.append( seedScramble );
-		seedHash.append( id );
+
+		if( id <= INT32_MAX && id >= INT_MIN )
+		{
+			// This branch shouldn't be needed, we'd like to just treat ids as 64 bit now ...
+			// but if we just took the branch below, that would changing the seeding of existing
+			// scenes.
+			seedHash.append( (int)id );
+		}
+		else
+		{
+			seedHash.append( id );
+		}
+
 		id = int( ( double( seedHash.h1() ) / double( UINT64_MAX ) ) * double( numSeeds ) );
 		id = id % numSeeds;  // For the rare case h1 / max == 1.0, make sure we stay in range
 	}
 	return id;
 }
-
-InternedString g_prototypeRootName( "root" );
-ConstInternedStringVectorDataPtr g_emptyNames = new InternedStringVectorData();
 
 }
 
@@ -329,6 +393,7 @@ class Instancer::EngineData : public Data
 			const std::string &position,
 			const std::string &orientation,
 			const std::string &scale,
+			const std::string &inactiveIds,
 			const std::string &attributes,
 			const std::string &attributePrefix,
 			const std::vector< PrototypeContextVariable > &prototypeContextVariables
@@ -337,13 +402,11 @@ class Instancer::EngineData : public Data
 				m_numPrototypes( 0 ),
 				m_numValidPrototypes( 0 ),
 				m_prototypeIndices( nullptr ),
-				m_ids( nullptr ),
 				m_positions( nullptr ),
 				m_orientations( nullptr ),
 				m_scales( nullptr ),
 				m_uniformScales( nullptr ),
-				m_prototypeContextVariables( prototypeContextVariables ),
-				m_hasDuplicates( false )
+				m_prototypeContextVariables( prototypeContextVariables )
 		{
 			if( !m_primitive )
 			{
@@ -352,13 +415,10 @@ class Instancer::EngineData : public Data
 
 			initPrototypes( mode, prototypeIndexName, rootsVariable, rootsList, prototypes );
 
-			if( const IntVectorData *ids = m_primitive->variableData<IntVectorData>( idName ) )
+			m_ids.initialize( m_primitive.get(), idName );
+			if( m_ids.size() && m_ids.size() != numPoints() )
 			{
-				m_ids = &ids->readable();
-				if( m_ids->size() != numPoints() )
-				{
-					throw IECore::Exception( fmt::format( "Id primitive variable \"{}\" has incorrect size", idName ) );
-				}
+				throw IECore::Exception( fmt::format( "Id primitive variable \"{}\" has incorrect size", idName ) );
 			}
 
 			if( const V3fVectorData *p = m_primitive->variableData<V3fVectorData>( position ) )
@@ -396,24 +456,118 @@ class Instancer::EngineData : public Data
 				}
 			}
 
-			if( m_ids )
+			if( m_ids.size() )
 			{
 				for( size_t i = 0, e = numPoints(); i < e; ++i )
 				{
-					int id = (*m_ids)[i];
+					int64_t id = m_ids.element(i);
 					auto ins = m_idsToPointIndices.try_emplace( id, i );
 					if( !ins.second )
 					{
+						// We have multiple indices trying to use this id.
 						if( !omitDuplicateIds )
 						{
 							throw IECore::Exception( fmt::format( "Instance id \"{}\" is duplicated at index {} and {}. This probably indicates invalid source data, if you want to hack around it, you can set \"omitDuplicateIds\".", id, m_idsToPointIndices[id], i ) );
 						}
 
-						// Invalidate the existing entry in the m_idsToPointIndices map - since we can't assign
-						// a consistent point index to this id, we omit all point indices that try to use this
-						// id
-						ins.first->second = std::numeric_limits<size_t>::max();
-						m_hasDuplicates = true;
+						if( !m_indicesInactive.size() )
+						{
+							m_indicesInactive.resize( numPoints(), false );
+						}
+
+						// If we're omitting duplicate ids, then we need to omit both the current index, and
+						// the index that first tried to use this id.
+						m_indicesInactive[ i ] = true;
+						m_indicesInactive[ ins.first->second ] = true;
+					}
+				}
+			}
+
+			std::vector<std::string> inactiveIdVarNames;
+			IECore::StringAlgo::tokenize( inactiveIds, ' ', inactiveIdVarNames );
+			for( std::string &inactiveIdVarName : inactiveIdVarNames )
+			{
+				if( m_primitive->variables.find( inactiveIdVarName ) == m_primitive->variables.end() )
+				{
+					continue;
+				}
+
+				const PrimitiveVariable *vertexInactiveVar = findVertexVariable( m_primitive.get(), inactiveIdVarName );
+				if( vertexInactiveVar )
+				{
+					if( const auto *vertexInactiveData = IECore::runTimeCast<BoolVectorData>( vertexInactiveVar->data.get() ) )
+					{
+						const std::vector<bool> &vertexInactive = vertexInactiveData->readable();
+						if( vertexInactive.size() != numPoints() )
+						{
+							throw IECore::Exception( fmt::format( "Inactive primitive variable \"{}\" has incorrect size", inactiveIdVarName ) );
+						}
+
+						if( !m_indicesInactive.size() )
+						{
+							// If we don't already have an inactive array set up, we can just directly copy the data
+							// from a vertex primitive variable. Technically, we might not even need to do this copy,
+							// if there aren't any other inactive vars we're merging with, we could just have a
+							// separate way of storing a const pointer for this case, but given that this data is
+							// 32X smaller than any of our other per-vertex data anyway, it's probably fine to pay
+							// the cost of copying it in exchange for slightly simpler code.
+							m_indicesInactive = vertexInactive;
+						}
+						else
+						{
+							for( unsigned int i = 0; i < vertexInactive.size(); i++ )
+							{
+								if( vertexInactive[i] )
+								{
+									m_indicesInactive[ i ] = true;
+								}
+							}
+						}
+
+						continue;
+					}
+				}
+
+				IdData idData;
+				idData.initialize( m_primitive.get(), inactiveIdVarName );
+
+				size_t idSize = idData.size();
+				if( !idSize )
+				{
+					continue;
+				}
+
+				if( !m_indicesInactive.size() )
+				{
+					m_indicesInactive.resize( numPoints(), false );
+				}
+
+				if( m_idsToPointIndices.size() )
+				{
+					for( size_t i = 0; i < idSize; i++ )
+					{
+						auto it = m_idsToPointIndices.find( idData.element(i) );
+						if( it == m_idsToPointIndices.end() )
+						{
+							// I wish I could throw here ... it would be a really helpful clue to get an error
+							// if you've accidentally chosen a bad id. But ids might be changing over time, so
+							// we probably need to allow someone to deactivate an id even if it doesn't exist
+							// on all frames.
+							continue;
+						}
+						m_indicesInactive[ it->second ] = true;
+					}
+				}
+				else
+				{
+					for( size_t i = 0; i < idSize; i++ )
+					{
+						int64_t id = idData.element(i);
+						if( id < 0 || id >= (int64_t)m_indicesInactive.size() )
+						{
+							continue;
+						}
+						m_indicesInactive[ id ] = true;
 					}
 				}
 			}
@@ -441,16 +595,16 @@ class Instancer::EngineData : public Data
 			return m_primitive ? m_primitive->variableSize( PrimitiveVariable::Vertex ) : 0;
 		}
 
-		size_t instanceId( size_t pointIndex ) const
+		int64_t instanceId( size_t pointIndex ) const
 		{
-			return m_ids ? (*m_ids)[pointIndex] : pointIndex;
+			return m_ids.size() ? m_ids.element( pointIndex ) : pointIndex;
 		}
 
-		size_t pointIndex( size_t i ) const
+		size_t pointIndex( int64_t i ) const
 		{
-			if( !m_ids )
+			if( !m_ids.size() )
 			{
-				if( i >= numPoints() )
+				if( i >= (int64_t)numPoints() || i < 0 )
 				{
 					throw IECore::Exception( fmt::format( "Instance id \"{}\" is invalid, instancer produces only {} children. Topology may have changed during shutter.", i, numPoints() ) );
 				}
@@ -483,14 +637,12 @@ class Instancer::EngineData : public Data
 				return -1;
 			}
 
-			if( m_hasDuplicates )
+			if( m_indicesInactive.size() )
 			{
 				// If there are duplicates in the id list, then some point indices will be omitted - we
-				// need to check each point index to see if it got assigned an id correctly
+				// need to check each point index to see if it got assigned an id correctly // TODO
 
-				int id = (*m_ids)[pointIndex];
-
-				if( m_idsToPointIndices.at(id) != pointIndex )
+				if( m_indicesInactive[pointIndex] )
 				{
 					return -1;
 				}
@@ -632,7 +784,7 @@ class Instancer::EngineData : public Data
 
 				if( v.seedMode )
 				{
-					scope.setAllocated( v.name, seedForPoint( pointIndex, v.primVar, v.numSeeds, v.seedScramble ) );
+					scope.setAllocated( v.name, seedForPoint( pointIndex, m_ids, v.numSeeds, v.seedScramble ) );
 					continue;
 				}
 
@@ -660,7 +812,7 @@ class Instancer::EngineData : public Data
 		{
 			if( v.seedMode )
 			{
-				result.append( seedForPoint( pointIndex, v.primVar, v.numSeeds, v.seedScramble ) );
+				result.append( seedForPoint( pointIndex, m_ids, v.numSeeds, v.seedScramble ) );
 				return;
 			}
 
@@ -915,13 +1067,13 @@ class Instancer::EngineData : public Data
 		std::vector<int> m_prototypeIndexRemap;
 		std::vector<int> m_prototypeIndicesAlloc;
 		const std::vector<int> *m_prototypeIndices;
-		const std::vector<int> *m_ids;
+		IdData m_ids;
 		const std::vector<Imath::V3f> *m_positions;
 		const std::vector<Imath::Quatf> *m_orientations;
 		const std::vector<Imath::V3f> *m_scales;
 		const std::vector<float> *m_uniformScales;
 
-		using IdsToPointIndices = std::unordered_map <int, size_t>;
+		using IdsToPointIndices = std::unordered_map <int64_t, size_t>;
 		IdsToPointIndices m_idsToPointIndices;
 
 		boost::container::flat_map<InternedString, AttributeCreator> m_attributeCreators;
@@ -929,7 +1081,7 @@ class Instancer::EngineData : public Data
 
 		const std::vector< PrototypeContextVariable > m_prototypeContextVariables;
 
-		bool m_hasDuplicates;
+		std::vector<bool> m_indicesInactive;
 
 		friend Instancer::EngineSplitPrototypesData;
 };
@@ -972,7 +1124,7 @@ class Instancer::EngineSplitPrototypesData : public Data
 				pointIndicesForPrototypeIndex[ constantPrototypeIndex ].reserve( m_engineData->numPoints() );
 			}
 
-			if( constantPrototypeIndex != -1 && !m_engineData->m_hasDuplicates )
+			if( constantPrototypeIndex != -1 && !m_engineData->m_indicesInactive.size() )
 			{
 				// If there's a single prototype, and no indices are being omitted because they are duplicates,
 				// then the list of point indices for the prototype is just an identity map of all integers
@@ -1200,6 +1352,7 @@ Instancer::Instancer( const std::string &name )
 	addChild( new StringPlug( "position", Plug::In, "P" ) );
 	addChild( new StringPlug( "orientation", Plug::In ) );
 	addChild( new StringPlug( "scale", Plug::In ) );
+	addChild( new StringPlug( "inactiveIds", Plug::In, "inactiveIds invisibleIds" ) );
 	addChild( new StringPlug( "attributes", Plug::In ) );
 	addChild( new StringPlug( "attributePrefix", Plug::In ) );
 	addChild( new BoolPlug( "encapsulate", Plug::In ) );
@@ -1341,154 +1494,164 @@ const Gaffer::StringPlug *Instancer::scalePlug() const
 	return getChild<StringPlug>( g_firstPlugIndex + 10 );
 }
 
-Gaffer::StringPlug *Instancer::attributesPlug()
+Gaffer::StringPlug *Instancer::inactiveIdsPlug()
 {
 	return getChild<StringPlug>( g_firstPlugIndex + 11 );
+}
+
+const Gaffer::StringPlug *Instancer::inactiveIdsPlug() const
+{
+	return getChild<StringPlug>( g_firstPlugIndex + 11 );
+}
+
+Gaffer::StringPlug *Instancer::attributesPlug()
+{
+	return getChild<StringPlug>( g_firstPlugIndex + 12 );
 }
 
 const Gaffer::StringPlug *Instancer::attributesPlug() const
 {
-	return getChild<StringPlug>( g_firstPlugIndex + 11 );
+	return getChild<StringPlug>( g_firstPlugIndex + 12 );
 }
 
 Gaffer::StringPlug *Instancer::attributePrefixPlug()
 {
-	return getChild<StringPlug>( g_firstPlugIndex + 12 );
+	return getChild<StringPlug>( g_firstPlugIndex + 13 );
 }
 
 const Gaffer::StringPlug *Instancer::attributePrefixPlug() const
 {
-	return getChild<StringPlug>( g_firstPlugIndex + 12 );
+	return getChild<StringPlug>( g_firstPlugIndex + 13 );
 }
 
 Gaffer::BoolPlug *Instancer::encapsulatePlug()
 {
-	return getChild<BoolPlug>( g_firstPlugIndex + 13 );
+	return getChild<BoolPlug>( g_firstPlugIndex + 14 );
 }
 
 const Gaffer::BoolPlug *Instancer::encapsulatePlug() const
 {
-	return getChild<BoolPlug>( g_firstPlugIndex + 13 );
+	return getChild<BoolPlug>( g_firstPlugIndex + 14 );
 }
 
 Gaffer::BoolPlug *Instancer::seedEnabledPlug()
 {
-	return getChild<BoolPlug>( g_firstPlugIndex + 14 );
+	return getChild<BoolPlug>( g_firstPlugIndex + 15 );
 }
 
 const Gaffer::BoolPlug *Instancer::seedEnabledPlug() const
 {
-	return getChild<BoolPlug>( g_firstPlugIndex + 14 );
+	return getChild<BoolPlug>( g_firstPlugIndex + 15 );
 }
 
 Gaffer::StringPlug *Instancer::seedVariablePlug()
 {
-	return getChild<StringPlug>( g_firstPlugIndex + 15 );
+	return getChild<StringPlug>( g_firstPlugIndex + 16 );
 }
 
 const Gaffer::StringPlug *Instancer::seedVariablePlug() const
 {
-	return getChild<StringPlug>( g_firstPlugIndex + 15 );
+	return getChild<StringPlug>( g_firstPlugIndex + 16 );
 }
 
 Gaffer::IntPlug *Instancer::seedsPlug()
 {
-	return getChild<IntPlug>( g_firstPlugIndex + 16 );
+	return getChild<IntPlug>( g_firstPlugIndex + 17 );
 }
 
 const Gaffer::IntPlug *Instancer::seedsPlug() const
 {
-	return getChild<IntPlug>( g_firstPlugIndex + 16 );
+	return getChild<IntPlug>( g_firstPlugIndex + 17 );
 }
 
 Gaffer::IntPlug *Instancer::seedPermutationPlug()
 {
-	return getChild<IntPlug>( g_firstPlugIndex + 17 );
+	return getChild<IntPlug>( g_firstPlugIndex + 18 );
 }
 
 const Gaffer::IntPlug *Instancer::seedPermutationPlug() const
 {
-	return getChild<IntPlug>( g_firstPlugIndex + 17 );
+	return getChild<IntPlug>( g_firstPlugIndex + 18 );
 }
 
 Gaffer::BoolPlug *Instancer::rawSeedPlug()
 {
-	return getChild<BoolPlug>( g_firstPlugIndex + 18 );
+	return getChild<BoolPlug>( g_firstPlugIndex + 19 );
 }
 
 const Gaffer::BoolPlug *Instancer::rawSeedPlug() const
 {
-	return getChild<BoolPlug>( g_firstPlugIndex + 18 );
+	return getChild<BoolPlug>( g_firstPlugIndex + 19 );
 }
 
 Gaffer::ValuePlug *Instancer::contextVariablesPlug()
 {
-	return getChild<ValuePlug>( g_firstPlugIndex + 19 );
+	return getChild<ValuePlug>( g_firstPlugIndex + 20 );
 }
 
 const Gaffer::ValuePlug *Instancer::contextVariablesPlug() const
 {
-	return getChild<ValuePlug>( g_firstPlugIndex + 19 );
+	return getChild<ValuePlug>( g_firstPlugIndex + 20 );
 }
 
 GafferScene::Instancer::ContextVariablePlug *Instancer::timeOffsetPlug()
 {
-	return getChild<ContextVariablePlug>( g_firstPlugIndex + 20 );
+	return getChild<ContextVariablePlug>( g_firstPlugIndex + 21 );
 }
 
 const GafferScene::Instancer::ContextVariablePlug *Instancer::timeOffsetPlug() const
 {
-	return getChild<ContextVariablePlug>( g_firstPlugIndex + 20 );
+	return getChild<ContextVariablePlug>( g_firstPlugIndex + 21 );
 }
 
 Gaffer::AtomicCompoundDataPlug *Instancer::variationsPlug()
 {
-	return getChild<AtomicCompoundDataPlug>( g_firstPlugIndex + 21 );
+	return getChild<AtomicCompoundDataPlug>( g_firstPlugIndex + 22 );
 }
 
 const Gaffer::AtomicCompoundDataPlug *Instancer::variationsPlug() const
 {
-	return getChild<AtomicCompoundDataPlug>( g_firstPlugIndex + 21 );
+	return getChild<AtomicCompoundDataPlug>( g_firstPlugIndex + 22 );
 }
 
 Gaffer::ObjectPlug *Instancer::enginePlug()
 {
-	return getChild<ObjectPlug>( g_firstPlugIndex + 22 );
+	return getChild<ObjectPlug>( g_firstPlugIndex + 23 );
 }
 
 const Gaffer::ObjectPlug *Instancer::enginePlug() const
 {
-	return getChild<ObjectPlug>( g_firstPlugIndex + 22 );
+	return getChild<ObjectPlug>( g_firstPlugIndex + 23 );
 }
 
 Gaffer::ObjectPlug *Instancer::engineSplitPrototypesPlug()
 {
-	return getChild<ObjectPlug>( g_firstPlugIndex + 23 );
+	return getChild<ObjectPlug>( g_firstPlugIndex + 24 );
 }
 
 const Gaffer::ObjectPlug *Instancer::engineSplitPrototypesPlug() const
 {
-	return getChild<ObjectPlug>( g_firstPlugIndex + 23 );
+	return getChild<ObjectPlug>( g_firstPlugIndex + 24 );
 }
 
 GafferScene::ScenePlug *Instancer::capsuleScenePlug()
 {
-	return getChild<ScenePlug>( g_firstPlugIndex + 24 );
+	return getChild<ScenePlug>( g_firstPlugIndex + 25 );
 }
 
 const GafferScene::ScenePlug *Instancer::capsuleScenePlug() const
 {
-	return getChild<ScenePlug>( g_firstPlugIndex + 24 );
+	return getChild<ScenePlug>( g_firstPlugIndex + 25 );
 }
 
 Gaffer::PathMatcherDataPlug *Instancer::setCollaboratePlug()
 {
-	return getChild<PathMatcherDataPlug>( g_firstPlugIndex + 25 );
+	return getChild<PathMatcherDataPlug>( g_firstPlugIndex + 26 );
 }
 
 const Gaffer::PathMatcherDataPlug *Instancer::setCollaboratePlug() const
 {
-	return getChild<PathMatcherDataPlug>( g_firstPlugIndex + 25 );
+	return getChild<PathMatcherDataPlug>( g_firstPlugIndex + 26 );
 }
 
 void Instancer::affects( const Plug *input, AffectedPlugsContainer &outputs ) const
@@ -1508,6 +1671,7 @@ void Instancer::affects( const Plug *input, AffectedPlugsContainer &outputs ) co
 		input == positionPlug() ||
 		input == orientationPlug() ||
 		input == scalePlug() ||
+		input == inactiveIdsPlug() ||
 		input == attributesPlug() ||
 		input == attributePrefixPlug() ||
 		input == seedEnabledPlug() ||
@@ -1603,6 +1767,7 @@ void Instancer::hash( const Gaffer::ValuePlug *output, const Gaffer::Context *co
 		positionPlug()->hash( h );
 		orientationPlug()->hash( h );
 		scalePlug()->hash( h );
+		inactiveIdsPlug()->hash( h );
 		attributesPlug()->hash( h );
 		attributePrefixPlug()->hash( h );
 		encapsulatePlug()->hash( h );
@@ -1695,7 +1860,7 @@ void Instancer::hash( const Gaffer::ValuePlug *output, const Gaffer::Context *co
 					for( size_t i = r.begin(); i != r.end(); ++i )
 					{
 						const size_t pointIndex = pointIndicesForPrototype[i];
-						size_t instanceId = engine->instanceId( pointIndex );
+						int64_t instanceId = engine->instanceId( pointIndex );
 						engine->setPrototypeContextVariables( pointIndex, scope );
 						IECore::MurmurHash instanceH;
 						instanceH.append( instanceId );
@@ -1791,15 +1956,12 @@ void Instancer::compute( Gaffer::ValuePlug *output, const Gaffer::Context *conte
 
 			if( seedContextName != "" )
 			{
-				const PrimitiveVariable *idPrimVar = findVertexVariable( primitive.get(), idPlug()->getValue() );
-				if( idPrimVar && idPrimVar->data->typeId() != IntVectorDataTypeId )
-				{
-					idPrimVar = nullptr;
-				}
-
 				int seeds = rawSeedPlug()->getValue() ? 0 : seedsPlug()->getValue();
 				int seedScramble = seedPermutationPlug()->getValue();
-				prototypeContextVariables.push_back( { seedContextName, idPrimVar, 0, false, true, seeds, seedScramble } );
+
+				// We set seedMode to true here, which means rather than reading a given primvar, this context
+				// variable will be driven by whatever is driving id.
+				prototypeContextVariables.push_back( { seedContextName, nullptr, 0, false, true, seeds, seedScramble } );
 			}
 
 			if( timeOffsetEnabled )
@@ -1833,6 +1995,7 @@ void Instancer::compute( Gaffer::ValuePlug *output, const Gaffer::Context *conte
 				positionPlug()->getValue(),
 				orientationPlug()->getValue(),
 				scalePlug()->getValue(),
+				inactiveIdsPlug()->getValue(),
 				attributesPlug()->getValue(),
 				attributePrefixPlug()->getValue(),
 				prototypeContextVariables
@@ -1948,7 +2111,7 @@ void Instancer::compute( Gaffer::ValuePlug *output, const Gaffer::Context *conte
 					for( size_t i = r.begin(); i != r.end(); ++i )
 					{
 						const size_t pointIndex = pointIndicesForPrototype[i];
-						size_t instanceId = engine->instanceId( pointIndex );
+						int64_t instanceId = engine->instanceId( pointIndex );
 						engine->setPrototypeContextVariables( pointIndex, scope );
 						ConstPathMatcherDataPtr instanceSet = prototypesPlug()->setPlug()->getValue();
 						PathMatcher pointInstanceSet = instanceSet->readable().subTree( *prototypeRoot );
@@ -2411,7 +2574,7 @@ IECore::ConstInternedStringVectorDataPtr Instancer::computeBranchChildNames( con
 		// the ids, not the point indices, and must be sorted. So we need to allocate a
 		// temp buffer of integer ids, before converting to strings.
 
-		std::vector<int> ids;
+		std::vector<int64_t> ids;
 		ids.reserve( pointIndicesForPrototype.size() );
 
 		const EngineData *engineData = esp->engine();
@@ -2427,7 +2590,7 @@ IECore::ConstInternedStringVectorDataPtr Instancer::computeBranchChildNames( con
 		InternedStringVectorDataPtr childNamesData = new InternedStringVectorData;
 		std::vector<InternedString> &childNames = childNamesData->writable();
 		childNames.reserve( ids.size() );
-		for( size_t id : ids )
+		for( int64_t id : ids )
 		{
 			childNames.emplace_back( id );
 		}
@@ -2986,7 +3149,7 @@ void Instancer::InstancerCapsule::render( IECoreScenePreview::Renderer *renderer
 					attribs = proto->m_rendererAttributes.get();
 				}
 
-				int instanceId = engines[0]->instanceId( pointIndex );
+				int64_t instanceId = engines[0]->instanceId( pointIndex );
 
 
 				if( !namePrefixLengths[protoIndex] )
@@ -3007,7 +3170,7 @@ void Instancer::InstancerCapsule::render( IECoreScenePreview::Renderer *renderer
 				// up being named when they use the non-encapsulated hierarchy.
 				std::string &name = names[ protoIndex ];
 				const int prefixLen = namePrefixLengths[ protoIndex ];
-				name.resize( namePrefixLengths[protoIndex] + std::numeric_limits< int >::digits10 + 1 );
+				name.resize( namePrefixLengths[protoIndex] + std::numeric_limits< int64_t >::digits10 + 1 );
 				name.resize( std::to_chars( &name[prefixLen], &(*name.end()), instanceId ).ptr - &name[0] );
 
 				IECoreScenePreview::Renderer::ObjectInterfacePtr objectInterface;