+++ Changes.md
@@ -1,11 +1,21 @@
 1.x.x.x (relative to 1.5.0.0a3)
 =======
 
+Features
+--------
+
+- Instancer : Added `inactiveIds` plug for selecting primitive variables to disable some instances.
+
 Improvements
 ------------
 
 - Light Editor : Added `is_sphere` column for Cycles lights.
 - Windows : Gaffer now uses the TBB memory allocator for significantly better performance.
+
+Fixes
+-----
+
+- Instancer : Added support 64 bit ints for ids ( matching what is loaded from USD ).
 
 Breaking Changes
 ----------------